+++ src/converter.ts
@@ -10,7 +10,7 @@ import {ConverterRuleToken} from './converterRule';
 
 /* NOT FOR BROWSER */
 
-import {html, font} from '../util/html';
+import {html} from '../util/html';
 import {Shadow} from '../util/debug';
 import {classes} from '../util/constants';
 import {flagsParent} from '../mixin/flagsParent';
@@ -139,13 +139,10 @@ export abstract class ConverterToken extends Token {
 			nocc = opt?.nocc,
 			{childNodes: [, ...rules]} = this;
 		if (nocc || flags.has('R') || this.getVariantFlags().size > 0) {
-			return font(this, (nocc ? '-{' : '') + html(rules, ';', opt) + (nocc ? '}-' : ''));
+			return (nocc ? '-{' : '') + html(rules, ';', opt) + (nocc ? '}-' : '');
 		} else if (flags.has('S')) {
-			return font(
-				this,
-				rules.find(({variant}) => variant)?.lastChild.toHtmlInternal(opt).trim()
-				?? rules[0].lastChild.toHtmlInternal(opt),
-			);
+			return rules.find(({variant}) => variant)?.lastChild.toHtmlInternal(opt).trim()
+				?? rules[0].lastChild.toHtmlInternal(opt);
 		}
 		return '';
 	}