+++ src/extLink.ts
@@ -14,7 +14,7 @@ import type {LintError} from '../base';
 /* NOT FOR BROWSER */
 
 import {normalizeSpace} from '../util/string';
-import {html, font} from '../util/html';
+import {html} from '../util/html';
 import {Shadow} from '../util/debug';
 import {magicLinkParent} from '../mixin/magicLinkParent';
 import type {MagicLinkParentBase} from '../mixin/magicLinkParent';
@@ -205,7 +205,6 @@ export abstract class ExtLinkToken extends Token {
 		const {length, lastChild} = this;
 		let trail = '',
 			innerText: string,
-			autonumber = false,
 			href: string | undefined;
 		if (length > 1) {
 			lastChild.normalize();
@@ -222,15 +221,13 @@ export abstract class ExtLinkToken extends Token {
 			innerText = lastChild.toHtmlInternal(opt);
 		} else {
 			({innerText} = this);
-			autonumber = true;
 		}
 		try {
 			({href} = this.getUrl());
 		} catch {}
-		const a = `<a rel="nofollow" class="external"${
+		return `<a rel="nofollow" class="external"${
 			href === undefined ? '' : ` href="${href}"`
 		}>${innerText}</a>${trail}`;
-		return autonumber ? font(this, a) : a;
 	}
 }