+++ src/magicLink.ts
@@ -21,7 +21,6 @@ import type {
 
 import {Shadow} from '../util/debug';
 import {classes} from '../util/constants';
-import {font} from '../util/html';
 import {syntax} from '../mixin/syntax';
 import type {SyntaxBase} from '../mixin/syntax';
 
@@ -273,14 +272,11 @@ export abstract class MagicLinkToken extends Token {
 		try {
 			url = this.getUrl();
 		} catch {}
-		return font(
-			this,
-			`<a${
-				type === 'magic-link' && protocol === 'ISBN'
-					? ''
-					: ` rel="nofollow" class="external${type === 'free-ext-link' ? ' free' : ''}"`
-			}${url === undefined ? '' : ` href="${typeof url === 'string' ? url : url.href}"`}>${innerText}</a>`,
-		);
+		return `<a${
+			type === 'magic-link' && protocol === 'ISBN'
+				? ''
+				: ` rel="nofollow" class="external${type === 'free-ext-link' ? ' free' : ''}"`
+		}${url === undefined ? '' : ` href="${typeof url === 'string' ? url : url.href}"`}>${innerText}</a>`;
 	}
 }